Move number formatting code into format-numbers.js
src/components/shared/MarkerTooltipContents.js:
src/utils/format-numbers.js:
    Move formatBytes and formatValueTotal into format-numbers.js.
